pyproject.toml
[tool.poetry]
name = "reqradar"
version = "0.2.0"
description = "需求透视 - 垂直领域需求分析 Agent"
readme = "README.md"
license = "MIT"
authors = ["easy <easy@example.com>"]
packages = [{ include = "reqradar", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
pyyaml = "^6.0.1"
pydantic = "^2.5.0"
structlog = "^24.1.0"
rich = "^13.7.0"
jinja2 = "^3.1.3"
chromadb = "^0.4.22"
sentence-transformers = "^2.3.1"
openai = "^1.12.0"
httpx = "^0.26.0"
gitpython = "^3.1.40"
numpy = "^1.26.0"
pdfplumber = {version = "^0.10.0", optional = true}
python-docx = {version = "^1.1.0", optional = true}

[tool.poetry.extras]
pdf = ["pdfplumber"]
docx = ["python-docx"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true

[tool.poetry.scripts]
reqradar = "reqradar.cli.main:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
